# BigDecimal: Replace ==ZERO by isZero, ~40% time reduction on main methods